Move symbols sync-schema to separate file
